move transactions table (#1278)
* wip: initial changes to move table

* chore: remove unused component

* Revert "chore: remove unused component"

This reverts commit 0db71a15538b776c73b752a98d2e825d890d2ea1.

* chore: remove unused component

* chore: use translation file

* fix: add missing p tags
