+++ src/App.tsx
@@ -28,7 +28,6 @@ import { FeatureFlags, useFeatureFlag } from './utils/hooks/use-feature-flag';
 const Bridge = React.lazy(() => import(/* webpackChunkName: 'bridge' */ '@/pages/Bridge'));
 const Strategies = React.lazy(() => import(/* webpackChunkName: 'strategies' */ '@/pages/Strategies'));
 const Transfer = React.lazy(() => import(/* webpackChunkName: 'transfer' */ '@/pages/Transfer'));
-const Transactions = React.lazy(() => import(/* webpackChunkName: 'transactions' */ '@/pages/Transactions'));
 const TX = React.lazy(() => import(/* webpackChunkName: 'tx' */ '@/pages/TX'));
 const Staking = React.lazy(() => import(/* webpackChunkName: 'staking' */ '@/pages/Staking'));
 const Dashboard = React.lazy(() => import(/* webpackChunkName: 'dashboard' */ '@/pages/Dashboard'));
@@ -181,9 +180,6 @@ const App = (): JSX.Element => {
                   </Route>
                   <Route path={PAGES.STAKING}>
                     <Staking />
-                  </Route>
-                  <Route path={PAGES.TRANSACTIONS}>
-                    <Transactions />
                   </Route>
                   <Route path={PAGES.TX}>
                     <TX />
