+++ src/pages/Bridge/IssueForm/SubmittedIssueRequestModal/index.tsx
@@ -36,7 +36,7 @@ const SubmittedIssueRequestModal = ({
       <ModalFooter>
         <InterlayRouterLink
           to={{
-            pathname: PAGES.TRANSACTIONS,
+            pathname: PAGES.BRIDGE,
             search: queryString.stringify({
               [QUERY_PARAMETERS.ISSUE_REQUEST_ID]: request.id
             })
