+++ src/utils/constants/general.ts
@@ -1,3 +1,3 @@
-const TABLE_PAGE_LIMIT = 20;
+const TABLE_PAGE_LIMIT = 5;
 
 export { TABLE_PAGE_LIMIT };
